Remove int casting on cy calc
Former-commit-id: 8d1ee68d8e05916cc4986932c660830ed1b61182

+++ GreenVision_Distance_WIP.py
@@ -79,7 +79,8 @@ while True:
     for c in ncontours:
         M = cv2.moments(c)
         if M["m00"] != 0:
-            cy = int(M["m01"] / M["m00"])
+            # cy = int(M["m01"] / M["m00"])
+            cy = M["m01"] / M["m00"]
         else:
             cy = 0, 0
         print('cy: {}'.format(cy))
